pyproject.toml
[tool.poetry]
name = "historia"
version = "0.1.0"
description = "Knowledge management system"
authors = ["Ole Kristian Losvik <losvik@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
wagtail = "^5.0.2"
Django = "^4.2.3"
wagtail-markdown = "^0.11.0"
django-storages = "^1.13.2"
boto3 = "^1.28.5"
dj-database-url = "^2.0.0"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
psycopg2-binary = "^2.9.6"
gunicorn = "^21.1.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
python-dotenv = {extras = ["cli"], version = "^1.0.0"}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
